requirements.txt
asgiref==3.5.2
asttokens==2.2.1
backcall==0.2.0
backports.zoneinfo==0.2.1
colorlog==6.7.0
decorator==5.1.1
dj-database-url==1.0.0
Django==4.1.1
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environ==1.0
executing==1.2.0
factory-boy==3.2.1
Faker==18.2.0
gunicorn==20.1.0
ipython==8.12.0
jedi==0.18.2
matplotlib-inline==0.1.6
mysqlclient==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.5.0
wcwidth==0.2.6
whitenoise==6.2.0
